requirements.txt
pyspark==3.3.0
delta-spark==2.2.0
pandas==1.5.0
numpy==1.23.3
pyarrow==10.0.0
PyYAML==6.0
sqlalchemy==1.4.41
requests>=2.25.0
findspark==1.3.0